Move ViewportOffset to its own file
There aren't any interdependencies with the rest of sliver.dart, and that file
is already quite complex.
